Added custom hook, updated VideoCards
