+++ src/pages/Home/Home.page.jsx
@@ -1,19 +1,16 @@
 import React, { useRef, useState } from 'react';
 import { Link, useHistory } from 'react-router-dom';
 import  Header  from '../../components/Header';
-import youtube from '../../apis/youtube';
 import VideoCards from '../../components/VideoCards';
 import { useAuth } from '../../providers/Auth';
 import './Home.styles.css';
 
 
 function HomePage() {
-  const [search, setSearch] = useState("");
-  const [videos, setVideos] = useState([]);
+  const [search, setSearch] = useState('');
 
   const updateSearch = (childData) => {
     setSearch(childData);
-    handleSubmit(childData);
   }
 
   const history = useHistory();
@@ -24,16 +21,6 @@ function HomePage() {
     event.preventDefault();
     logout();
     history.push('/');
-  }
-
-  const handleSubmit = async (search) => {
-    const response = await youtube.get('/search', {
-      params: {
-        q: search
-      }
-    })
-    console.log(response.data);
-    setVideos(response.data.items);
   }
 
   return (
@@ -57,7 +44,7 @@ function HomePage() {
             {search === "" ? (
               <h2>Look for something!</h2>
             ) : (
-              <VideoCards videos={videos} />
+              <VideoCards query={search} />
             )}
           </>
         )}
